Added single enemy to level 1
Also added prefabs and scripts to add to each level, level 1 loads hallway scene, level 2 loads volcano scene...
zombie prefab has same move script as the drone, so probably wont look good on level 2 or 3, drone would look great on level 3, just attach the movetowardslev3 script to the drone and a collider that is a sphere. or you could change the reload scene to increase the time.